python_scripts/checkin/requirements.txt
requests
PyExecJS
pytz
ddddocr
